Cost Calculation (Switch conditional)
The purpose of this project was to enhance my proficiency in building operators in PHP, while also utilizing HTML and CSS. The project involved creating a cost calculation system for switches, aiming to provide users with a comprehensive summary of length units. Its primary focus was on individuals who require accurate fund calculations for household development. Through this project, I aimed to improve my understanding of cost calculations and further develop my skills in PHP programming with web technologies